Renamed package and moved to application module.
