+++ src/main/java/org/ednovo/gooru/application/client/newhome/IsNewHomeView.java
@@ -25,7 +25,7 @@
 /**
  *
 */
-package org.ednovo.gooru.client.mvp.test;
+package org.ednovo.gooru.application.client.newhome;
 
 import org.ednovo.gooru.application.client.gin.IsViewWithHandlers;
 
@@ -45,7 +45,7 @@ import org.ednovo.gooru.application.client.gin.IsViewWithHandlers;
  *
  * @Reviewer:
  */
-public interface IsTestView extends IsViewWithHandlers<TestUiHandlers>{
+public interface IsNewHomeView extends IsViewWithHandlers<NewHomeUiHandlers>{
 
 	/**
 	 *
