+++ src/main/java/org/ednovo/gooru/application/client/newhome/NewHomeView.ui.xml
